+++ out-of-the-box/design-toolkit.esm.bundled.js
@@ -361,6 +361,17 @@ EchoButton.styles = [
       .button--ghost:hover:not(:disabled) {
         background-color: var(--context-color-alpha);
       }
+
+      .button--soft {
+        background-color: var(--context-color-alpha);
+        color: var(--context-color);
+        border: none;
+      }
+
+      .button--soft:hover:not(:disabled) {
+        background-color: var(--context-color);
+        color: white;
+      }
     `,
 ];
 __decorate([