+++ src/types/index.ts
@@ -1,4 +1,4 @@
-export type EchoButtonVariant = 'default' | 'link' | 'outline' | 'ghost';
+export type EchoButtonVariant = 'default' | 'link' | 'outline' | 'ghost' | 'soft';
 export type EchoSize = 'xs' | 'small' | 'medium' | 'large';
 export type EchoContext =
   | 'danger'